separate reporter file
